Checkpoint: made hapi support asynchronous template engines
This required restructuring view Manager.render() and _get() to return
their results asynchronously.

Added two new flags to the views section of server options:
asyncCompile -- set to true if the template engine compiles
asynchronously
asyncRender -- set to true if the template engine renders asynchronously

Remaining work:
- test the with-layout case
- test an async engine other than blade
- verify that synchronous template engines still work as desired
- write new automated tests
- make sure the changes look like everything else in the project

The provided tests fail with a zlib assertion failure, but that test
fails in the master branch as well. Not sure if this causes testing to
abort with some tests un-run.