Made layouts work with different template engines.
Added three unit tests for async template engines: valid, invalid, and
valid with layout. Fixed the bugs revealed in my dry-coded layout
rendering.

To make blade layouts work, added a new option to the "views" options
section. Set `layoutFile` to the name of the layout file you wish to
use, with suffix appropriate for your layout engine. The default is
"layout.html", so existing hapi layout code continues to work.

This isn't generalized enough, since you'll want to be able to override
layouts on the fly for some pages, but it's a start.